+++ sql/role-records.sql
@@ -1,1 +1,5 @@
-INSERT INTO `role` (`id`, `name`, `deleted`, `updated_at`, `created_at`) VALUES (1, 'guest', 0, now(), now()), (2, 'member', 0, now(), now()), (3, 'admin', 0, now(), now()), (4, 'superadmin', 0, now(), now());
+INSERT INTO `role` (`id`, `name`, `deleted`, `hidden`, `updated_at`, `created_at`) VALUES
+(1,	'guest',	0,	0,	NOW(),	NOW()),
+(2,	'member',	0,	0,	NOW(),	NOW()),
+(3,	'admin',	0,	0,	NOW(),	NOW()),
+(4,	'superadmin',	0,	0,	NOW(),	NOW());